Add code to stop possible local pow mining operation
The inserted function mining_status_ändern() will be called when another
node has alreay comleted the search for the next proof of work. In case
that a local pow searching operation is running, it will be stopped. In
this way the local node won't be running a pow searching operation for a
block, where is has already been found.