Add initial sync with masternode
A function has been added that initially syncs the client blockchain
with the masternode blockchain. This way when starting up the local
blockchain the client will be up to date. There has also been added
a function that checks if the masternode blockchain is valid. Code
for the slightly different masternode blockchain and api has also
been added. When syncing, the new transaction that have not yet
been put into a block will also by copied over to the client after
the validity of the masternode blockchain has been checked.