Add more syncing functions
The client will now send his pow to the masternode which
then verifies it and if verification is successful client
node and masternode will create new block. This will be
changes in future updates to work with block time. Furthermore
the client sends new transactions to the masternode where they
will be added to the masternodes unconfirmed transaction list.
After a positive http reply the client node will also add this
transaction to his list for unconfirmed transactions.