Add check if pow has already been calculated
Before mining/calculating the new pow the local node will check if
pow has already been calculated by another node through the a function
that calls the /mining/status endpoint on the masternode. If it returns
False, the client will calculate the next pow and go on with the rest
of the mining function. Else it will return http status 200 with a
message saying the pow has already been mined. In the first case
when the masternode gets the pow it will set a variable called
already_mined to True. This value will stay like that until a new
block is created, then it will be changes to False again.