Add /set/name endpoint to client node
Client node will now have to set the clients name before interacting
with the masternode in a way that requires the clients name e.g.
sending transactions, requesting own transactions, requesting
current balance. This name will also be used when mining and sending
the pow to the masternode, for the block reward. So no name is
anymore needed when calling /spezifische/transaktionen or
/spezifische/balance and no sender is anymore needed when calling
/transaktionen/neu.