Add block halving and balance check
Before client node can send a transaction it's balance is checked.
If it is smaller than the transaction value the client wants to
send the transaction fails. If it is bigger or equal the transaction
goes through. Furthermore block halving has been implemented into
the masternode. After a certain amount of blocks has been reached
(first halving at 7200, then 7200*2=x, then x*2 etc.) the block
reward will be halved, so from the initial 500 to 250, then 125 etc.