Add burning mechanism & genesis block transactions
With the implemented coin burning mechanism a deflation effect is
achieved. From all the registered users 10% are randomly chosen.
If a selected users balance is equal to zero a new user is be
chosen. Then the summed up balance of all the selected users is
caculated and compared with the value of 0.25% of the ledgers
balance. If the summed up balance is greater than 0.25% of the
ledgers balance, the 0.25% of the ledgers balance is set as
to_be_burned value. Else 10% of the summed up balance is set as
the to_be_burned value. Then the to_be_burned value is distributed
equally between the selected users, depending on how much balance
they have. After that the calculated amount is burned from their
wallets and the whole to_be_burned amount is burned from the
ledgers balance. The 0.25% are subject to change. Furthermore a
function in the masternode_blockchain.py file has been created
where a starting amount of users are created and a starting
balance is added to their accounts. The starting balance each
user gets is random, but lies in a selected range which is also
subject to change.