Add sync of new master nodes
New master nodes will initiate a startup procededure when started
where they use lookup servers to check if there is active master
nodes in the network. If there is, the master node will choose a
random one out of the network and sync from it. If there is none,
the masternode will generate a genesis block and continue normally
from there on.