Add early implementation of complete network sync
Masternodes now share the same blockchain. New blocks are created
when a miner discovers a new block. The mining difficulty has been
increased from finding a proof that generates a hash with 4 leading
zeros to finding a proof that generates a hash with 7 leading zeros.
Current problems are blockchain forks at a specific point, when two
miners each find valid proof at the same time, but these two proofs
are not the same. This way two miners update their masternode on the
new proof and the masternodes update the network. The other masternodes
in the network will accepts whichever proof reaches them first and
decline the proof that reaches them next. This way the blockchain is
forked, but both blockchains are valid. This problem needs to be
fixed in future commits.